+++ src/pages/Diagnostico/Diagnostico.js
