+++ src/pages/Sobre/Sobre.css
@@ -1,0 +1,37 @@
+.retanguloStyle{
+    width: '9px';
+    height: '188px';
+    flex-shrink: 0;
+    background-color: '#0B2A45';
+};
+
+.tituloStyle{
+    font-size: '25px';
+    text-align: 'center';
+    margin-left: '20px';
+};
+
+.tituloContainerStyle{
+    display: 'flex';
+    align-items: 'center';
+};
+
+.paragrafoStyle{
+    margin-top: '40px';
+    margin-left: '40px';
+    margin-bottom: '100px';
+    font-size: '20px';
+};
+
+.imagemStyle{
+    position: 'relative';
+};
+
+.sobreNosTextoStyle{
+    position: 'absolute';
+    bottom: '40px'; 
+    left: '30px'; 
+    font-size: '40px';
+    font-weight: 'bold';
+    color: 'white';
+};
